# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Artemis Application Server API
  contact: {name: Stephan Krusche, email: krusche@tum.de}
  version: "10.0"
servers:
- {url: 'http://localhost:8080', description: Generated server url}
...
paths:
...
  /api/hyperion/programming-exercises/{exerciseId}/generation/variants:
    post:
      tags: [hyperion-programming-variant]
      operationId: generateProgrammingVariant
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VariantGenerationRequest'}
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationJobStart'}
...
  /api/hyperion/exercises/{exerciseId}/generate-variant:
    post:
      tags: [hyperion-exercise-variant]
      operationId: generateVariant
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VariantGenerationRequest'}
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VariantJobStart'}
...
  /api/hyperion/variant-jobs/{jobId}:
    get:
      tags: [hyperion-exercise-variant]
      operationId: getJobDetail
      parameters:
      - name: jobId
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VariantJobDetail'}
    delete:
      tags: [hyperion-exercise-variant]
      operationId: cancelJob
      parameters:
      - name: jobId
        in: path
        required: true
        schema: {type: string}
      responses:
        '200': {description: OK}
...
  /api/hyperion/admin/exercises/{exerciseId}/hyperion-wedged-slot:
    get:
      tags: [admin-hyperion-generation]
      operationId: getWedgedSlot
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationWedgedSlot'}
...
  /api/hyperion/admin/exercises/{exerciseId}/hyperion-wedged-slots/{token}:
    delete:
      tags: [admin-hyperion-generation]
      operationId: recoverWedgedSlot
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      - name: token
        in: path
        required: true
        schema: {type: string}
      - name: reason
        in: query
        required: true
        schema: {type: string}
      responses:
        '200': {description: OK}
...
components:
  schemas:
    VariantJob:
      type: object
      properties:
        jobId: {type: string}
        sourceExerciseId: {type: integer, format: int64}
        courseId: {type: integer, format: int64}
        sourceExerciseTitle: {type: string}
        exerciseType:
          type: string
          enum: [text, programming, modeling, file-upload, quiz]
        phase:
          type: string
          enum: [ANALYZING, PLANNING, PROVISIONING, TRANSFORMING, VERIFYING, REPAIRING,
            FINALIZING, COMPLETED, DRAFT_WITH_WARNINGS, FAILED, CANCELLED]
        failedInPhase:
          type: string
          enum: [ANALYZING, PLANNING, PROVISIONING, TRANSFORMING, VERIFYING, REPAIRING,
            FINALIZING, COMPLETED, DRAFT_WITH_WARNINGS, FAILED, CANCELLED]
        failureDetail: {type: string}
        instructorSummary: {type: string}
        attempt: {type: integer, format: int32}
        maxAttempts: {type: integer, format: int32}
        variantExerciseId: {type: integer, format: int64}
        variantExerciseTitle: {type: string}
        request: {$ref: '#/components/schemas/VariantGenerationRequest'}
        warnings:
          type: array
          items: {type: string}
        totalTokensUsed: {type: integer, format: int64}
        startedAt: {type: string, format: date-time}
        finishedAt: {type: string, format: date-time}
...
    ExerciseGenerationJobStart:
      type: object
      properties:
        jobId: {type: string}
        exerciseId: {type: integer, format: int64}
        sourceExerciseId: {type: integer, format: int64}
      required: [jobId]
...
    VariantJobStart:
      type: object
      properties:
        jobId: {type: string}
...
    VariantJobDetail:
      type: object
      properties:
        job: {$ref: '#/components/schemas/VariantJob'}
        stepOutputs:
          type: object
          additionalProperties:
            type: array
            items: {$ref: '#/components/schemas/StepOutput'}
        request: {$ref: '#/components/schemas/VariantGenerationRequest'}
...
    ExerciseGenerationWedgedSlot:
      type: object
      properties:
        exerciseId: {type: integer, format: int64}
        token: {type: string}
        kind: {type: string}
        ownerNodeId: {type: string}
        startedAt: {type: string, format: date-time}
        ownerLeftCluster: {type: boolean}
...
    VariantGenerationRequest:
      type: object
      properties:
        targetDifficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        domainText: {type: string, maxLength: 2000, minLength: 0}
        narrativeStyle:
          type: string
          enum: [TECHNICAL, REALISTIC, CREATIVE]
        additionalInstructions: {type: string, maxLength: 2000, minLength: 0}
        placement: {$ref: '#/components/schemas/VariantPlacement'}
...
    CreateExerciseVariantGroup:
      type: object
      properties:
        title: {type: string, maxLength: 255, minLength: 0}
        maxPoints: {type: number, format: double, minimum: 0}
        releaseDate: {type: string, format: date-time}
        startDate: {type: string, format: date-time}
        dueDate: {type: string, format: date-time}
        assessmentDueDate: {type: string, format: date-time}
        exampleSolutionPublicationDate: {type: string, format: date-time}
      required: [title]
...
    VariantPlacement:
      type: object
      properties:
        type:
          type: string
          enum: [EXISTING_GROUP, NEW_GROUP, STANDALONE, SAME_EXAM_GROUP]
        existingGroupId: {type: integer, format: int64}
        newGroup: {$ref: '#/components/schemas/CreateExerciseVariantGroup'}
...
    StepOutput:
      type: object
      properties:
        summary: {type: string}
        detail: {type: string}
...