# src/main/webapp/app/openapi/model/variant-job-detail.ts
/**
 * Artemis Application Server API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API Version: 10.0
 * Contact: krusche@tum.de
 *
 * NOTE: This file is auto-generated. Do not edit manually.
 */

import type { VariantJob } from './variant-job';
import type { StepOutput } from './step-output';
import type { VariantGenerationRequest } from './variant-generation-request';

export interface VariantJobDetail {
    job?: VariantJob;
    stepOutputs?: { [key: string]: Array<StepOutput> };
    request?: VariantGenerationRequest;
}